+++ requests_magic/spider.py
@@ -2,6 +2,7 @@ from typing import Callable
 from .request import Request
 from .item import Item
 import requests
+import warnings
 
 
 class Spider:
@@ -13,11 +14,11 @@ class Spider:
 
     def parse(self, result: requests.Response, request: Request):
         """
-        Default parse function, Can return/yield requests/items, or lists containing them
+        example to parse function, Can return/yield requests/items, or lists containing them
         """
         pass
 
-    def request(self, url: str, data: dict = None, callback: Callable = None,
+    def request(self, url: str, callback: Callable, data: dict = None,
                 meta: dict = None, headers: dict = None) -> Request:
         """
         Create a request
@@ -28,11 +29,12 @@ class Spider:
         :param callback: parse function
         :param meta: extra information
         """
+        warnings.warn("Recommend direct instantiation Request", DeprecationWarning)
         if meta is None:
             meta = {}
         if headers is None:
             headers = {}
-        return Request(url=url, spider=self,
+        return Request(url=url,
                        callback=callback if callback else self.parse,
                        data=data, meta=meta, headers=headers)
 
@@ -40,9 +42,10 @@ class Spider:
         """
         Create a item
         (it is not recommended to instantiate the Item class directly, Using this method is the best)
-        :param data: realy data
+        :param data: data
         :param meta: extra information
         """
+        warnings.warn("Recommend direct instantiation Item", DeprecationWarning)
         if meta is None:
             meta = {}
-        return Item(data, spider=self, meta=meta)
+        return Item(data, meta=meta)